requirements.txt
groq
google-generativeai
langchain
langchain-core
langchain-community
python-dotenv
langdetect
pymupdf
python-docx
fastapi
uvicorn
streamlit
